+++ api/models/terminology_search_results.py
@@ -1,13 +1,13 @@
 from api.models.no_null_camel_model import NoNullCamelModel
 from api.database.db_omop_tables import Concept
-from typing import List
+from typing import List, Optional
 from api.features.terminologysearch.system_map import reverse_lookup_system
 
 class ConceptResult(NoNullCamelModel):
     name: str
     code: str
     system: str
-    has_presets: bool = False
+    has_presets: Optional[bool] = None
 
     @classmethod
     def from_concept_table(cls, concept: Concept, has_presets: bool = False):