+++ api/utilities/settings.py
@@ -10,7 +10,7 @@ class Settings(BaseSettings):
     cors_origins: str = "*"
     omop_database_url: str | None = None
     show_fhirsheets_logs: bool = False
-    enable_run_logs: bool = False # TODO Implement
+    enable_run_logs: bool = False
     iteration_limit: int = 100 # Max number of iterations for event generation.
     root_path: str | None = None
     force_reseed: bool = False